+++ quince/ui/common/avatar.py
