+++ quince/ui/game_frame_factory.py
@@ -2,7 +2,7 @@
     Factory object that gets used to inject dependencies
     into the GameApp constructor.
 """
-from quince.ui.components.game_frame import GameFrame
+from quince.ui.game_frame import GameFrame
 
 
 class GameFrameFactory(object):
