+++ quince/ui/opponents/hand.py
