+++ quince/ui/start_gui.py
@@ -1,12 +1,11 @@
 #!/usr/bin/env python
 
+"""Graphical application initialization
 """
-    Graphical application initialization
-"""
-from quince.ui.components.game_app import GameApp
-from quince.ui.components.about import AboutFactory
-from quince.ui.components.top_menu.top_menu import TopMenuFactory
-from quince.ui.components.score_report.score_report import ScoreReport
+from quince.ui.game_app import GameApp
+from quince.ui.about import AboutFactory
+from quince.ui.top_menu.top_menu import TopMenuFactory
+from quince.ui.score_report.score_report import ScoreReport
 
 
 def start():
