+++ quince/ui/top_menu/avatar_picker.py
